+++ Riot/AppDelegate.h
@@ -26,6 +26,7 @@
 #import "Analytics.h"
 
 #import "ThemeService.h"
+#import "UniversalLink.h"
 
 #pragma mark - Notifications
 /**
@@ -41,6 +42,11 @@ extern NSString *const kAppDelegateNetworkStatusDidChangeNotification;
 extern NSString *const AppDelegateDidValidateEmailNotification;
 extern NSString *const AppDelegateDidValidateEmailNotificationSIDKey;
 extern NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey;
+
+/**
+ Posted when the property 'lastHandledUniversalLink' has changed. Notification object and userInfo will be nil.
+ */
+extern NSString *const AppDelegateUniversalLinkDidChangeNotification;
 
 @interface AppDelegate : UIResponder <UIApplicationDelegate, MXKCallViewControllerDelegate, UISplitViewControllerDelegate, UINavigationControllerDelegate, JitsiViewControllerDelegate, UNUserNotificationCenterDelegate>
 {
@@ -86,6 +92,11 @@ extern NSString *const AppDelegateDidValidateEmailNotificationClientSecretKey;
 
 // Current selected room id. nil if no room is presently visible.
 @property (strong, nonatomic) NSString *visibleRoomId;
+
+/**
+ Last handled universal link (url will be formatted for several hash keys).
+ */
+@property (nonatomic, readonly) UniversalLink *lastHandledUniversalLink;
 
 // New message sound id.
 @property (nonatomic, readonly) SystemSoundID messageSound;