+++ CHANGES.rst
@@ -7,6 +7,7 @@ Improvements:
  * Self-verification: Do not display incoming self verification requests at the top of the Complete Security screen.
  * Verification: Do not talk about QR code if only emoji is possible (#3035).
  * Registration: Prefill email field when opened with universal link (PR #3173).
+ * Complete Security: Support SAS verification start (#3183).
 
 Bug fix:
  * AuthenticationViewController: Remove fallback to matrix.org when authentication failed (PR #3165).